Extract PlayerIcon into separate class
